+++ src/main/java/com/nova/cls/data/services/criteria/goods/GroupIdCriterion.java
@@ -3,7 +3,7 @@ package com.nova.cls.data.services.criteria.goods;
 public class GroupIdCriterion extends GoodsCriterion {
     private static final String SQL = "groupId = ?";
 
-    public GroupIdCriterion(Integer groupId) {
+    public GroupIdCriterion(long groupId) {
         super(SQL, groupId);
     }
 }
