+++ src/main/java/com/nova/cls/data/services/criteria/goods/MaxPriceCriterion.java
@@ -3,7 +3,7 @@ package com.nova.cls.data.services.criteria.goods;
 public class MaxPriceCriterion extends GoodsCriterion {
     private static final String SQL = "price <= ?";
 
-    public MaxPriceCriterion(int price) {
+    public MaxPriceCriterion(long price) {
         super(SQL, price);
     }
 }
