+++ src/main/java/com/nova/cls/data/services/criteria/goods/MinPriceCriterion.java
@@ -3,7 +3,7 @@ package com.nova.cls.data.services.criteria.goods;
 public class MinPriceCriterion extends GoodsCriterion {
     private static final String SQL = "price >= ?";
 
-    public MinPriceCriterion(int price) {
+    public MinPriceCriterion(long price) {
         super(SQL, price);
     }
 }
